[api] Move getXpos function from utilities to api package
