+++ packages/api/package.json
@@ -22,7 +22,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@broad/utilities": "*",
     "compression": "^1.7.1",
     "core-js": "^2.5.0",
     "cors": "^2.8.1",
